Rename output files when we find duplicates
This is possible because archipelago supports adding `{NUMBER}` to your
player name which would automatically generate incremental player names.
Unfortunately when I implemented that special case in the unicity check
for yamls in the room I kinda forgot about the zip file. Which meant
that a player uploading a yaml with the same name twice would generate a
zip file with the same file name twice.

I didn't even know zip supported that but thankfully it does because
that would have been a pain to debug why someone wouldn't have had one
of their world generated. Instead it asked me what I wanted to do with
the duped file when unzipping which immediately uncovered the bug.

Just adding a incremental suffix to duplicated file names fixes that.