Add some `api/` routes that would return proper statuses
When writing the route to download a YAML, I didn't take into account
that I might have to call it through some code. Which means the route
would take you back to the index, display a nice error message but still
return a 200.

This commit adds a new error type as well as an associated result, which
implements another type of responder, which will use the provided status
if it has one, and default to 500. With that, we can display an error to
the user if they try to view a YAML file in a room while the file has
already been deleted instead of showing them the index page HTML.

Fixes #12