Add an env variable to allow skipping apworld updates
When developing locally, I don't need them to update every time I
restart the lobby. I usually just comment it, but as proven by the
previous commit, I always forget to uncomment it. Just add an env
variable, which is here by default to skip updating the apworlds on
launch.