Use the right timezone offset for the date entered by a user
If you create a room for a timezone that isn't yet here (i.e
before/after DST), you end up sending the offset for your current
timezone instead of the destination one, shifting the close date by one
hour compared to what it should be.

We fix this by using the date provided by the user instead to get the
timezone offset.