Switch YAML validation to work queues
This is still very early work and not ready for production but it's a
good start. This commit changes how YAML validation works, going from a
synchronous `lobby -> validator` to a work queue system where the
validation jobs are put on a queue and picked up by independent workers.

Of course right now the only way of creating validation jobs is on
upload which means that in the end we're just making this more complex
as we're putting the job on the queue and then waiting for it to be
resolved.