Rework YAML validation and the YAML table
This includes two major pieces of work that went well together.

First, YAML validation is now non-definitive, it can now be updated. It
is currently being updated in 2 cases. If the room saves while open and
the apworlds used for the original validation don't match. When the
index updates, if the room is open then for each YAML, if the new
apworlds don't match the ones it was validated with. There's also an API
route although it is currently unused and might disappear.

Second, the YAML table got a bit of a rework. Because I added more
information to it, I thought that now was a good time to add proper
filtering and sorting to it. Honestly, should've done that separately
but here we are...