Update how flash errors are displayed
Because cookies have a size limit set to 4kB, some error messages were
overflowing that... Which lead to nginx returning a 502 for those
requests. Instead of storing flash msgs in the cookie, we now store a
uuid and store errors in redis instead, tied to that uuid.

Because of `Context` being required in `session`, I cracked and made the
whole thing a lib with a main reexported in main.rs which is why this
diff is horrible.

Fixes #55