Display a message if the provided session cookie isn't valid
Technically we should be checking for this on a regular basis but I feel
dirty enough doing it once on startup... And in practice we use this
tool once for 8h and then restart it for the next one since the whole
config is in env variables so unless the session cookie expires in the
middle (please no), it should never matter. And if it ends up being a
problem then I'll check on a schedule I guess.

I tried finding a way to get responses out of commands but there is
none. The "API" they provide if you can call it that is not great and
basically returns a redirect whatever happens and then the command is
executed asynchronously. They somewhat work around that in the frontend
by polling the logs aggressively when you emit a command and stopping
that aggressive poll on the first text that comes up so if it's an item
you're waiting another 10s... Maybe we could do something by parsing the
logs but that's way too brittle as they've shown in the past that
they'll change the log format silently. So yeah, this is a best effort
to fix #73 for commands emitted through the web form (basically
anything except `release`)

Closes #73 even though I'm not happy with the result, but there's
nothing else I can do without hosting AP...