Simplify room settings
Originally I wrote this to avoid duping the settings struct. It ended up
nightmarish but I never bothered to roll that back. Now with a macro
with can avoid the duplication and remove the horrendous code I had to
make stuff Queryable and use diesel's derives instead. Of course it's
not exactly perfect because we do end up with two different structs
anyway, but they're guaranteed to always be in sync since they're genned
by the same code and we only have one conversion function. Adding a
setting won't be as painful as it used to be