Don't return empty bundles from `get_bundle_by_id`
A `YamlBundle` assumes that it will always contain at least one YAML.
Because we're querying the list of YAMLs for a given bundle ID, querying
a bundle that doesn't exist would result in an empty one. Later on if we
did any operation on said bundle (like getting the owner id for it when
deleting it for example), it would panic because we'd assume the bundle
isn't empty.

Fixes sentry issue AP-LOBBY-A