Make diff parsing a lot faster by windowing comparisons
Also ignore lines that have a wildly different length in comparison.
There are cases where an apworld contains files with 10k lines where
each of these get a tiny change and that makes the best line matches
very unhappy leading to egregious loading times. This fixes it by
windowing the blocks in which we compare things to. Sure it might miss
some matches in large files but it makes some diff go from "I'm never
going to load" to loading in 2 seconds so I'll take it.

Fixes #95